shift.js
shifts funcionando correctamente (FINALMENTE)
